+++ CONTRIBUTING.md
@@ -32,6 +32,24 @@ Crash and force-close reports must include a crash log or relevant device consol
 
 Describe the problem, use case, proposed solution, and alternatives. Wait for explicit maintainer approval before starting implementation. An open or popular feature request is not approval by itself.
 
+## Project generation
+
+The Xcode project is generated by [Tuist](https://tuist.dev) and is not tracked in git, so there are no `project.pbxproj` diffs or merge conflicts to resolve.
+
+```bash
+mise install                          # installs the pinned Tuist version
+cd tvosApp && tuist generate          # writes and opens NuvioTV.xcworkspace
+```
+
+When your change touches the project rather than the code:
+
+- Edit [tvosApp/Project.swift](./tvosApp/Project.swift), never the generated `.xcodeproj`.
+- Adding, moving, or deleting a Swift file needs **no** project edit — sources are globbed. Just re-run `tuist generate`.
+- Re-run `tuist generate` after pulling, in case the manifest changed.
+- Do not commit `tvosApp/NuvioTV.xcodeproj` or `tvosApp/NuvioTV.xcworkspace`; both are ignored.
+
+For on-device signing, copy `tvosApp/Configurations/Signing.local.example.xcconfig` to `Signing.local.xcconfig` and set your own Team ID. That file is untracked, and the project generates and builds unsigned without it. See [tvosApp/Configurations/README.md](./tvosApp/Configurations/README.md).
+
 ## tvOS testing notes
 
 - Test on a physical Apple TV when possible; the simulator cannot play AV1.
