+++ README.md
@@ -125,10 +125,12 @@ Contributions are welcome. See [CONTRIBUTING.md](./CONTRIBUTING.md) for contribu
 
 - macOS with Xcode installed.
 - Apple TV simulator runtime installed in Xcode.
-- CocoaPods if `tvosApp/Pods` needs to be regenerated.
+- [Tuist](https://tuist.dev) to generate the Xcode project. Install it with [mise](https://mise.jdx.dev): `mise install` picks up the pinned version from [mise.toml](./mise.toml).
 - Network access for catalog metadata, source lookups, and Swift Package resolution.
 
-The Xcode project targets Apple TV (`SDKROOT = appletvos`) with bundle id `com.nuvio.app.tv`. The tvOS deployment target is configured in [project.pbxproj](./tvosApp/NuvioTV.xcodeproj/project.pbxproj).
+The Xcode project is **generated by Tuist and is not tracked in git**. Edit [tvosApp/Project.swift](./tvosApp/Project.swift) rather than the `.xcodeproj`, then re-run `tuist generate`. Source files are picked up by glob, so adding a file needs no project edit at all.
+
+The app targets Apple TV (`SDKROOT = appletvos`) with bundle id `com.pyksel.nuviotvos` and a tvOS 17.5 deployment target, all declared in `Project.swift`.
 
 ## Setup
 
@@ -137,19 +139,15 @@ git clone <your-fork-url> NuvioTVOS
 cd NuvioTVOS
 ```
 
-Install pods if the CocoaPods workspace has not been generated:
+Install the pinned toolchain and generate the Xcode project:
 
 ```bash
+mise install
 cd tvosApp
-pod install
-cd ..
+tuist generate
 ```
 
-Open the tvOS workspace:
-
-```bash
-open tvosApp/NuvioTV.xcworkspace
-```
+`tuist generate` opens the workspace in Xcode. Use `--no-open` to only write it.
 
 Use the `NuvioTV` scheme and an Apple TV simulator.
 
@@ -166,6 +164,7 @@ If no Apple TV simulator is booted, open Simulator or Xcode first and start one,
 You can also build directly with Xcode:
 
 ```bash
+(cd tvosApp && tuist generate --no-open)
 xcodebuild \
   -workspace tvosApp/NuvioTV.xcworkspace \
   -scheme NuvioTV \
@@ -188,10 +187,9 @@ Catalogs and metadata use configurable catalog, playback, and subtitle endpoints
 
 ## Tests
 
-Unit and UI test targets live in:
+Unit tests live in [tvosApp/NuvioTVTests](./tvosApp/NuvioTVTests).
 
-- [tvosApp/NuvioTVTests](./tvosApp/NuvioTVTests)
-- [tvosApp/NuvioTVUITests](./tvosApp/NuvioTVUITests)
+[tvosApp/NuvioTVUITests](./tvosApp/NuvioTVUITests) contains UI test sources that no target currently builds; see [docs/tuist-migration.md](./docs/tuist-migration.md).
 
 Run tests from Xcode, or with:
