+++ scripts/build-ipa.sh
@@ -13,6 +13,17 @@ cleanup() {
 }
 trap cleanup EXIT
 
+# The Xcode project is generated by Tuist and is not tracked in git.
+echo "==> Generating Xcode project with Tuist..."
+if command -v tuist >/dev/null 2>&1; then
+  (cd "$ROOT_DIR/tvosApp" && tuist generate --no-open)
+elif command -v mise >/dev/null 2>&1; then
+  (cd "$ROOT_DIR/tvosApp" && mise exec -- tuist generate --no-open)
+else
+  echo "error: tuist not found. Install it with 'mise install' (see mise.toml)." >&2
+  exit 1
+fi
+
 echo "==> Archiving ${TVOS_SCHEME} for tvOS (Release, unsigned)..."
 xcodebuild archive \
   -workspace "$TVOS_WORKSPACE" \
