+++ scripts/run-mobile.sh
@@ -47,12 +47,18 @@ first_booted_tvos_simulator() {
     | sed -E 's/.*\(([A-F0-9-]+)\) \(Booted\).*/\1/'
 }
 
-ensure_tvos_pods() {
-  if [[ ! -f "$TVOS_APP_DIR/Pods/Manifest.lock" ]]; then
-    require_command pod
-
-    echo "Generating tvOS CocoaPods project..."
-    (cd "$TVOS_APP_DIR" && pod install)
+ensure_tvos_project() {
+  # The Xcode project is generated by Tuist and is not tracked in git.
+  if [[ ! -d "$TVOS_APP_DIR/NuvioTV.xcodeproj" ]]; then
+    echo "Generating tvOS Xcode project with Tuist..."
+  fi
+  if command -v tuist >/dev/null 2>&1; then
+    (cd "$TVOS_APP_DIR" && tuist generate --no-open)
+  elif command -v mise >/dev/null 2>&1; then
+    (cd "$TVOS_APP_DIR" && mise exec -- tuist generate --no-open)
+  else
+    echo "error: tuist not found. Install it with 'mise install' (see mise.toml)." >&2
+    exit 1
   fi
 }
 
@@ -191,7 +197,7 @@ run_ios_physical() {
 run_tvos_simulator() {
   require_command xcodebuild
   require_command xcrun
-  ensure_tvos_pods
+  ensure_tvos_project
 
   local simulator_id
   simulator_id="$(first_booted_tvos_simulator)"
