--- tvosApp/Podfile
@@ -1,27 +1,0 @@
-# Podfile for NuvioTV
-# Pure Swift/SwiftUI tvOS application
-
-platform :tvos, '15.1'
-
-target 'NuvioTV' do
-  # Use frameworks for Swift compatibility
-  use_frameworks!
-
-  # Add any native iOS/tvOS dependencies here
-  # Example: pod 'Alamofire', '~> 5.6'
-
-  # Future: Add pods for native features
-  # - Analytics/crash reporting (if needed)
-  # - Media playback enhancements
-  # - Networking utilities
-end
-
-post_install do |installer|
-  installer.pods_project.targets.each do |target|
-    target.build_configurations.each do |config|
-      # Set minimum deployment target
-      config.build_settings['TVOS_DEPLOYMENT_TARGET'] = '15.1'
-      config.build_settings['IPHONEOS_DEPLOYMENT_TARGET'] = '15.1'
-    end
-  end
-end
